refactor: check has its own file
