+++ src/__test__/geometry/check.test.ts
@@ -1,0 +1,4 @@
+import { isKingChecked } from '../../geometry/check.ts';
+describe('King is checked', () => {
+  it('proves that king is checked in bishop direction', () => {});
+});
