+++ src/geometry/types.ts
@@ -1,11 +1,13 @@
 export const Rows = [1, 2, 3, 4, 5, 6, 7, 8] as const;
 export const Columns = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h'] as const;
 export const FigureLetters = ['k', 'q', 'r', 'b', 'n', 'p', 'K', 'Q', 'R', 'B', 'N', 'P'] as const;
+export const CastlingLetters = ['K', 'Q', 'k', 'q'] as const;
 
 export type Row = (typeof Rows)[number];
 export type Column = (typeof Columns)[number];
 export type FigureLetter = (typeof FigureLetters)[number];
 export type EnPassentColumn = Column | '-';
+export type CastlingLetter = (typeof CastlingLetters)[number] | '-';
 
 export type Color = 'white' | 'black' | 'none';
 export type Figure = 'PAWN' | 'KNIGHT' | 'BISHOP' | 'ROOK' | 'KING' | 'QUEEN';
@@ -14,7 +16,13 @@ export type Position = {
   row: Row;
   column: Column;
 };
-export type Move = Position & { isTaken?: boolean; isTakenEnPassent?: boolean };
+export type MoveProperties = {
+  isTaken?: boolean;
+  isTakenEnPassent?: boolean;
+  isPromotion?: boolean;
+  castling?: CastlingLetter;
+};
+export type Move = Position & MoveProperties; //extra info for move
 export type Piece = {
   color: Color;
   figure: Figure;
